Fix 4277 PR-fast coverage lane
